# Conflicts:
#	tests/TestCase/Validation/ValidationTest.php

+++ src/Database/Log/LoggingStatement.php
@@ -24,7 +24,6 @@ use Exception;
  */
 class LoggingStatement extends StatementDecorator
 {
-
     /**
      * Logger instance responsible for actually doing the logging task
      *
@@ -78,7 +77,7 @@ class LoggingStatement extends StatementDecorator
      */
     protected function _log($query, $params, $startTime)
     {
-        $query->took = round((microtime(true) - $startTime) * 1000, 0);
+        $query->took = (int)round((microtime(true) - $startTime) * 1000, 0);
         $query->params = $params ?: $this->_compiledParams;
         $query->query = $this->queryString;
         $this->getLogger()->log($query);
